insight-weaver-main/backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
requests==2.31.0
reportlab==4.0.7
python-dotenv==1.0.0
PyPDF2==3.0.1
